feat: add report-to directive and set Reporting-Endpoints header (#22)
Add the report-to instruction to the Content-Security-Policy and the Content-Security-Policy-Report-Only headers.
The report-uri instruction has been deprecated in favor of the report-to but is not yet broadly supported so both instructions should be specified (in browsers that support report-to, the report-uri directive is ignored).
Note that the report-to instruction requires the endpoints to be defined with the Reporting-Endpoints header.

Also setup the Cypress tests and configure the Github workflows accordingly.